if t1 and t2 were the same the difference would 1e-17 and the test would fail
git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@93007 16731396-06f5-0310-8873-f7f720988828

+++ resources/test/leap_second_test.py
@@ -533,7 +533,7 @@ for i in range(10000):
         leap_seconds = -sum([ x>d1 and x<=d2 for x in leap_sec_mjd])
 
     #check that the delta is correct
-    assert  (t1-t2)/I3Units.second == (d1-d2)*86400 + leap_seconds
+    assert  int((t1-t2)/I3Units.second) == (d1-d2)*86400 + leap_seconds
 
     #check I3time + double is correct
     time_delta = ((d2-d1)*86400L- leap_seconds)*I3Units.second
